tensorflow_fastapi/requirement.txt
absl-py==0.14.1
anyio==3.3.4
asgiref==3.4.1
astunparse==1.6.3
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
clang==5.0
click==8.0.3
colorama==0.4.4
fastapi==0.70.0
flatbuffers==1.12
gast==0.4.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
h11==0.12.0
h5py==3.1.0
idna==3.3
Jinja2==3.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
Pillow==8.4.0
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
python-multipart==0.0.5
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
sniffio==1.2.0
starlette==0.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.7
uvicorn==0.15.0
Werkzeug==2.0.2
wrapt==1.12.1
